Add files via upload
This Python project is a comprehensive Health Management System designed for individuals and trainers with multiple clients. With this system, users can log their daily exercise routines or diet plans and retrieve them whenever needed. Additionally, the system allows users to manage and keep track of their clients' health data and progress, making it an ideal tool for fitness trainers or nutritionists. The program is user-friendly and easy to navigate, with separate options for logging and retrieving health data for each client. The system's primary goal is to help individuals and trainers monitor their health and wellness and take appropriate measures to achieve their fitness goals. Overall, this Health Management System is an efficient and effective way to manage personal or clients' health data, diet plans, and exercise routines. In short it is a easy and simple method and even beginner can easily understand the code!