Add files via upload
Health Management System is a simple Python project that allows users to log and retrieve their health data. The system is designed for three persons named "NS, JB, and CP" and provides two options for users to choose from - log or retrieve. Once the user selects an option, they can select from two categories - exercise or food. The system logs the user's data in a text file named after the person and category. Users can retrieve their data by selecting the retrieve option and specifying the person and category. This project is designed to be beginner-friendly and can be easily modified to add additional features and functionality.