Health Management System (Beginner friendly code)
The README file for the "Health Management System" repository provides an overview of the project and its features, along with instructions for getting started and using the system. It also includes information on contributing to the projectd. This README file helps users understand what the project does, how to use it, and how to get involved, making it a valuable resource for anyone interested in the "Health Management System" project.